requirements.txt
iniconfig==2.3.0
packaging==26.0
pluggy==1.6.0
polars==1.38.1
polars-runtime-32==1.38.1
Pygments==2.19.2
pytest==9.0.2
PyYAML==6.0.3
